Create the specific package for the Share-related components.
